+++ lib/src/binder/simple_binder.dart
@@ -33,18 +33,7 @@ class SimpleBinder<S extends BinderState> extends BasicBinder<S> {
         );
 
   @override
-  Widget bindState(BoundWidgetBuilder<S> builder) {
-    return StreamBuilder<S>(
-      stream: _uiStatePipe.stream.distinct(),
-      builder: (context, snapshot) {
-        if (snapshot.data == null) {
-          return const SizedBox.shrink();
-        } else {
-          return builder(context, snapshot.data!);
-        }
-      },
-    );
-  }
+  Stream<S> getStateStream() => _uiStatePipe.stream.distinct();
 
   @override
   void emitUiState(S state) {